FPLA-2208: Removal of hearingBookingService. Minor defintion fix
