+++ ccd-definition/AuthorisationCaseEvent/CareSupervision/DEPRECATED.json
@@ -591,11 +591,13 @@
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "hearingBookingDetails",
-    "AccessControl": [
-      {
-        "UserRoles": ["caseworker-publiclaw-courtadmin", "caseworker-publiclaw-gatekeeper", "caseworker-publiclaw-judiciary", "caseworker-publiclaw-magistrate", "caseworker-publiclaw-solicitor"],
-        "CRUD": "R"
-      }
-    ]
+    "UserRoles": [
+      "caseworker-publiclaw-courtadmin",
+      "caseworker-publiclaw-gatekeeper",
+      "caseworker-publiclaw-judiciary",
+      "caseworker-publiclaw-magistrate",
+      "caseworker-publiclaw-solicitor"
+    ],
+    "CRUD": "R"
   }
 ]
